Initial data and folders to save code results
This compressed file corresponds to 9 folders, 7 of them have the types of templates to be evaluated, within each one you can save as many Excel documents (representing the crosses) as you want.
There is another folder that has the reference data provided by the ministries of education for their validation.
And there is a last folder where the validation results are saved. Ques is the same Excel document that results in a validation report: when it was done, which sheets and rows were found, which ones were not.
And for each sheet the fields that did not meet the validation and why they do not meet it.